Added problem explanation config, updated readme
Similar to the last commit, except also made it optional for there to
be an explanation for a problem attempt. Also updated readme.md to
explain that such configuration can be set from the main.py file and
does not have to be done manually.

Moved the config dictionary to its own line for better programming
style.